[style]: show more button added to happening feed component
